File upload fix, with minor style changes on job-details.
Upload fix (temporary) :

added a new variable in upload.ts that stores the filepath, then made a rule of thumb - if the upload object's filepath is not null and its url is null, this means the file's been affected with Async issue.

Due to the async nature of the upload, we can't get the file upload url on the  job post page. If you ask firebase the file's download URL there, it'll throw a 404 error. So, you won't be able to download the attachment there.(hence the slight re-style).

By the time the client or the provider reaches the jobs page, the file would've already been firmly uploaded to firebase. So, on the page's init, we can request the file's now-available download URL. This will only occur if and only if the job's attached file has no url, and this will only occur once.